Many changes
Largest change is changing Matrix to Matrix4x4, changing fields of that
struct from 4 Vector4s to 16 floats, and updating all uses of it. Also
cleaned up some of the other mass classes, and the audio util

+++ Pencil.Gaming/Audio/ALUtils.cs
@@ -124,8 +124,8 @@ namespace Pencil.Gaming.Audio {
 				out ALFormat format,
 				out uint sampleRate,
 				out uint avgBytesPerSec,
-				out short bytesPerSample,
-				out short bitsPerSample) {
+				out ushort bytesPerSample,
+				out ushort bitsPerSample) {
 
 				LoadWavExt(File.ReadAllBytes(file), out data, out chunkSize, out format, out sampleRate, out avgBytesPerSec, out bytesPerSample, out bitsPerSample);
 			}
@@ -157,8 +157,8 @@ namespace Pencil.Gaming.Audio {
 				out ALFormat format,
 				out uint sampleRate,
 				out uint avgBytesPerSec,
-				out short bytesPerSample,
-				out short bitsPerSample) {
+				out ushort bytesPerSample,
+				out ushort bitsPerSample) {
 
 				using (MemoryStream ms = new MemoryStream()) {
 					if (!file.CanRead) {
@@ -181,7 +181,7 @@ namespace Pencil.Gaming.Audio {
 				out uint sampleRate) {
 
 				uint dummyui;
-				short dummys;
+				ushort dummys;
 				LoadWavExt(file, out data, out dummyui, out format, out sampleRate, out dummyui, out dummys, out dummys);
 			}
 
@@ -192,15 +192,15 @@ namespace Pencil.Gaming.Audio {
 				out ALFormat format,
 				out uint sampleRate,
 				out uint avgBytesPerSec,
-				out short bytesPerSample,
-				out short bitsPerSample) {
+				out ushort bytesPerSample,
+				out ushort bitsPerSample) {
 
 #if DEBUG
 				Stopwatch sw = new Stopwatch();
 				sw.Start();
 #endif
 
-				short channels;
+				ushort channels;
 
 				int ptrOffset = 4;
 				if (sound [0] != 'R' || sound [1] != 'I' || sound [2] != 'F' || sound [3] != 'F') {
@@ -218,15 +218,15 @@ namespace Pencil.Gaming.Audio {
 				//ptrOffset = 20;
 				//formatType = ((short)(((short)sound[1 + ptrOffset] << 8) | ((short)sound[0 + ptrOffset])));
 				ptrOffset = 22;
-				channels = (short)(((short)sound [1 + ptrOffset] << 8) | ((short)sound [0 + ptrOffset]));
+				channels = (ushort)((sound [1 + ptrOffset] << 8) | (sound [0 + ptrOffset]));
 				ptrOffset = 24;
 				sampleRate = ((uint)sound [3 + ptrOffset] << 24) | ((uint)sound [2 + ptrOffset] << 16) | ((uint)sound [1 + ptrOffset] << 8) | ((uint)sound [ptrOffset]);
 				ptrOffset = 28;
 				avgBytesPerSec = ((uint)sound [3 + ptrOffset] << 24) | ((uint)sound [2 + ptrOffset] << 16) | ((uint)sound [1 + ptrOffset] << 8) | ((uint)sound [ptrOffset]);
 				ptrOffset = 32;
-				bytesPerSample = (short)(((short)sound [1 + ptrOffset] << 8) | ((short)sound [0 + ptrOffset]));
+				bytesPerSample = (ushort)((sound [1 + ptrOffset] << 8) | (sound [0 + ptrOffset]));
 				ptrOffset = 34;
-				bitsPerSample = (short)(((short)sound [1 + ptrOffset] << 8) | ((short)sound [0 + ptrOffset]));
+				bitsPerSample = (ushort)((sound [1 + ptrOffset] << 8) | (sound [0 + ptrOffset]));
 				if (sound [36] != 'd' || sound [37] != 'a' || sound [38] != 't' || sound [39] != 'a') {
 					throw new Exception("Invalid file format.");
 				}
